fixed some conversion warnings
Former-commit-id: 04aff8c55eda75d9e122ef15af4bfb7e8a1f21a3
Former-commit-id: bc9f8697cd3c6f02c7195442aeee4da8d19b5880

+++ src/fswAlgorithms/imageProcessing/houghCircles/houghCircles.cpp
@@ -138,7 +138,7 @@ void HoughCircles::UpdateState(uint64_t CurrentSimNanos)
 
     circleBuffer.timeTag = this->sensorTimeTag;
     circleBuffer.cameraID = imageBuffer.cameraID;
-    for( size_t i = 0; i < this->expectedCircles && i<circles.size(); i++ )
+    for( int i = 0; i < this->expectedCircles && i<(int) circles.size(); i++ )
     {
         circleBuffer.circlesCenters[2*i] = circles[i][0];
         circleBuffer.circlesCenters[2*i+1] = circles[i][1];
